fix(#110): 새 여행 생성 후 편집 눌렀을 때 다른 여행계획 일정 노출되는 버그 해결
